Migrated Abstractions.Tests to use AutoFixture.NSubstitute instead of manually creating mocks, and changed file structure to single file per class
