--- src/Abstractions.Tests/Extensions/MultipleContentBuilderExtensions/MultipleContentBuilderExtensionsTests.cs
@@ -1,6 +1,0 @@
-﻿namespace TemplateFramework.Abstractions.Tests.Extensions;
-
-public partial class MultipleContentBuilderExtensionsTests
-{
-    protected IMultipleContentBuilder CreateSut() => Substitute.For<IMultipleContentBuilder>();
-}
