--- src/Abstractions.Tests/Extensions/TemplateContextExtensions/TemplateContextExtensionsTests.GetModelFromContextByType.cs
@@ -1,20 +1,0 @@
-﻿namespace TemplateFramework.Abstractions.Tests.Extensions.TemplateContextExtensions;
-
-public partial class TemplateContextExtensionsTests
-{
-    public class GetModelFromContextByType : TemplateContextExtensionsTests
-    {
-        [Fact]
-        public void Without_Arguments_Works_Correctly()
-        {
-            // Arrange
-            var sut = CreateSut();
-
-            // Act
-            sut.GetModelFromContextByType<string>();
-
-            // Assert
-            sut.Received().GetModelFromContextByType<string>(null);
-        }
-    }
-}
